Split CommonHost into focused partial files
Copilot-Session: 8ea243c0-fd25-4425-97ed-30fef1f98629
