+++ src/Platform/Microsoft.Testing.Platform/Hosts/CommonTestHost.cs
@@ -2,15 +2,11 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Microsoft.Testing.Platform.Capabilities.TestFramework;
-using Microsoft.Testing.Platform.Extensions;
 using Microsoft.Testing.Platform.Extensions.ArtifactPostProcessing;
-using Microsoft.Testing.Platform.Extensions.TestFramework;
 using Microsoft.Testing.Platform.Extensions.TestHost;
 using Microsoft.Testing.Platform.Helpers;
 using Microsoft.Testing.Platform.IPC;
 using Microsoft.Testing.Platform.Logging;
-using Microsoft.Testing.Platform.Messages;
-using Microsoft.Testing.Platform.OutputDevice;
 using Microsoft.Testing.Platform.ServerMode;
 using Microsoft.Testing.Platform.Services;
 using Microsoft.Testing.Platform.Telemetry;
@@ -22,7 +18,7 @@ namespace Microsoft.Testing.Platform.Hosts;
 /// This doesn't represent an orchestrator host.
 /// </summary>
 [StackTraceHidden]
-internal abstract class CommonHost(ServiceProvider serviceProvider) : IHost
+internal abstract partial class CommonHost(ServiceProvider serviceProvider) : IHost
 {
     public ServiceProvider ServiceProvider => serviceProvider;
 
@@ -262,318 +258,4 @@ internal abstract class CommonHost(ServiceProvider serviceProvider) : IHost
     }
 
     protected abstract Task<int> InternalRunAsync(CancellationToken cancellationToken);
-
-    protected static async Task ExecuteRequestAsync(ProxyOutputDevice outputDevice, ITestSessionContext testSessionInfo,
-        ServiceProvider serviceProvider, BaseMessageBus baseMessageBus, ITestFramework testFramework, TestHost.ClientInfo client)
-    {
-        // Reset the shared, application-scoped coverage accumulator at the start of every request here, in the
-        // common host/request lifecycle, so it happens for all output modes (terminal, pipe, server, custom)
-        // rather than only when the terminal device renders. Without this a prior session's coverage rows and
-        // thresholds would be reprinted and its threshold-failure verdict could poison a later session.
-        serviceProvider.GetRequiredService<TestCoverageResult>().Reset();
-
-        await DisplayBeforeSessionStartAsync(outputDevice, testSessionInfo).ConfigureAwait(false);
-        CancellationToken cancellationToken = testSessionInfo.CancellationToken;
-        try
-        {
-            try
-            {
-                IPlatformOpenTelemetryService? otelService = serviceProvider.GetPlatformOTelService();
-                using (otelService?.StartActivity("OnTestSessionStarting"))
-                {
-                    await NotifyTestSessionStartAsync(testSessionInfo, baseMessageBus, serviceProvider, otelService).ConfigureAwait(false);
-                }
-
-                using (otelService?.StartActivity("TestFrameworkInvoker"))
-                {
-                    await serviceProvider.GetTestFrameworkInvoker().ExecuteAsync(testFramework, client, cancellationToken).ConfigureAwait(false);
-                }
-
-                using (otelService?.StartActivity("OnTestSessionEnding"))
-                {
-                    await NotifyTestSessionEndAsync(testSessionInfo, baseMessageBus, serviceProvider, otelService).ConfigureAwait(false);
-                }
-            }
-            catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-            {
-                // Do nothing we're canceled
-            }
-
-            // We keep the display after session out of the OperationCanceledException catch because we want to notify the IPlatformOutputDevice
-            // also in case of cancellation. Most likely it needs to notify users that the session was canceled.
-            await DisplayAfterSessionEndRunAsync(outputDevice, testSessionInfo).ConfigureAwait(false);
-        }
-        finally
-        {
-            // The message bus shutdown handshake must complete before the services - and with them every
-            // IDataConsumer - get disposed, otherwise a consumer can still be inside ConsumeAsync while it is
-            // being disposed. NotifyTestSessionEndAsync does it on the happy path, but it is skipped whenever the
-            // session is canceled or fails, so we close the handshake here for every outcome.
-            await EnsureMessageBusDisabledAsync(baseMessageBus, serviceProvider).ConfigureAwait(false);
-        }
-    }
-
-    /// <summary>
-    /// Disables the message bus, which stops accepting new payloads and awaits every consumer loop, so that no
-    /// <see cref="Extensions.IDataConsumer.ConsumeAsync"/> can still be running once the consumers are disposed.
-    /// </summary>
-    /// <remarks>
-    /// This is a best-effort safety net that runs on teardown paths, including the ones that are already unwinding
-    /// because of a cancellation or a failure. Disabling is idempotent, so on the regular path this is a no-op and
-    /// any consumer failure has already been surfaced by the real call. Failures here are logged and swallowed
-    /// rather than replacing the exception that is being propagated; a consumer that survives the handshake is
-    /// reported through <see cref="BaseMessageBus.ConsumersStillRunning"/> and left undisposed.
-    /// </remarks>
-    protected static async Task EnsureMessageBusDisabledAsync(BaseMessageBus baseMessageBus, IServiceProvider serviceProvider)
-    {
-        try
-        {
-            await baseMessageBus.DisableAsync().ConfigureAwait(false);
-        }
-        catch (Exception ex)
-        {
-            await LogShutdownWarningAsync(serviceProvider, $"Failed to disable the message bus during shutdown: {ex}").ConfigureAwait(false);
-        }
-    }
-
-    private static async Task LogShutdownWarningAsync(IServiceProvider serviceProvider, string message)
-    {
-        try
-        {
-            if (serviceProvider.GetService<ILoggerFactory>()?.CreateLogger(nameof(CommonHost)) is { } logger)
-            {
-                await logger.LogWarningAsync(message).ConfigureAwait(false);
-            }
-        }
-        catch (Exception)
-        {
-            // This runs on teardown paths that are often already unwinding, and the logging stack is itself
-            // being torn down (a disposed LoggerFactoryProxy throws from CreateLogger, and providers can fail
-            // from LogAsync). A failed warning must not replace the exception we were reporting, nor abort the
-            // rest of the disposal.
-        }
-    }
-
-    private static async Task DisplayBeforeSessionStartAsync(ProxyOutputDevice outputDevice, ITestSessionContext sessionInfo)
-    {
-        // Display before session start
-        await outputDevice.DisplayBeforeSessionStartAsync(sessionInfo.CancellationToken).ConfigureAwait(false);
-
-        if (outputDevice.OriginalOutputDevice is ITestSessionLifetimeHandler testSessionLifetimeHandler)
-        {
-            await testSessionLifetimeHandler.OnTestSessionStartingAsync(sessionInfo).ConfigureAwait(false);
-        }
-    }
-
-    private static async Task DisplayAfterSessionEndRunAsync(ProxyOutputDevice outputDevice, ITestSessionContext sessionInfo)
-    {
-        // Display after session end even when the session cancellation token is canceled.
-        // We intentionally pass a non-cancelable token so final output/cleanup notifications are not skipped.
-        await outputDevice.DisplayAfterSessionEndRunAsync(CancellationToken.None).ConfigureAwait(false);
-
-        // We want to ensure that the output service is the last one to run
-        if (outputDevice.OriginalOutputDevice is ITestSessionLifetimeHandler testSessionLifetimeHandlerFinishing)
-        {
-            await testSessionLifetimeHandlerFinishing.OnTestSessionFinishingAsync(sessionInfo).ConfigureAwait(false);
-        }
-    }
-
-    private static async Task NotifyTestSessionStartAsync(ITestSessionContext testSessionContext, BaseMessageBus baseMessageBus, ServiceProvider serviceProvider, IPlatformOpenTelemetryService? otelService)
-    {
-        TestSessionLifetimeHandlersContainer? testSessionLifetimeHandlersContainer = serviceProvider.GetService<TestSessionLifetimeHandlersContainer>();
-        if (testSessionLifetimeHandlersContainer is null)
-        {
-            return;
-        }
-
-        foreach (ITestSessionLifetimeHandler testSessionLifetimeHandler in testSessionLifetimeHandlersContainer.TestSessionLifetimeHandlers)
-        {
-            using IPlatformActivity? activity = otelService?.StartActivity(testSessionLifetimeHandler.Uid, testSessionLifetimeHandler.ToOTelTags());
-            await testSessionLifetimeHandler.OnTestSessionStartingAsync(testSessionContext).ConfigureAwait(false);
-        }
-
-        // Drain messages generated by the session start notification before to start test execution.
-        await baseMessageBus.DrainDataAsync().ConfigureAwait(false);
-    }
-
-    private static async Task NotifyTestSessionEndAsync(ITestSessionContext testSessionContext, BaseMessageBus baseMessageBus, ServiceProvider serviceProvider, IPlatformOpenTelemetryService? otelService)
-    {
-        // Drain messages generated by the test session execution before to process the session end notification.
-        await baseMessageBus.DrainDataAsync().ConfigureAwait(false);
-
-        TestSessionLifetimeHandlersContainer? testSessionLifetimeHandlersContainer = serviceProvider.GetService<TestSessionLifetimeHandlersContainer>();
-        if (testSessionLifetimeHandlersContainer is null)
-        {
-            // Nothing else to notify. The bus is disabled by the caller's teardown safety net
-            // (EnsureMessageBusDisabledAsync), which runs for every outcome of the session.
-            return;
-        }
-
-        IShutdownProgressReporter? shutdownProgressReporter = serviceProvider.GetService<IShutdownProgressReporter>();
-
-        // First, we call OnTestSessionFinishingAsync on all non-consumers.
-        bool hasNonDataConsumers = false;
-        foreach (ITestSessionLifetimeHandler testSessionLifetimeHandler in testSessionLifetimeHandlersContainer.TestSessionLifetimeHandlers)
-        {
-            if (testSessionLifetimeHandler is IDataConsumer)
-            {
-                // At first, we don't call this for data consumers.
-                // We want to do this for potentially publishers-only handlers.
-                // The order here matters, because one handler can produce a message that is consumed by another.
-                // By making the consumers last, we reduce the likelihood of issues related to ordering.
-                // One case where this is important is the dependency between Code Coverage's ITestSessionLifetimeHandler and TRX's ITestSessionLifetimeHandler.
-                // We must run Code Coverage ITestSessionLifetimeHandler implementation first, as it will publish SessionFileArtifact.
-                // The SessionFileArtifact is expected to be consumed by TRX's implementation of ITestSessionLifetimeHandler.
-                // In that case, ITestSessionLifetimeHandler of CC is a producer-only handler which we will run first.
-                // Then we will drain the message bus to ensure TRX handler have consumed the SessionFileArtifact.
-                // Then we run TRX OnTestSessionFinishingAsync.
-                continue;
-            }
-
-            hasNonDataConsumers = true;
-
-            using (otelService?.StartActivity(testSessionLifetimeHandler.Uid, testSessionLifetimeHandler.ToOTelTags()))
-            using (shutdownProgressReporter?.Track(testSessionLifetimeHandler.Uid, testSessionLifetimeHandler.DisplayName, nameof(ITestSessionLifetimeHandler.OnTestSessionFinishingAsync)))
-            {
-                await testSessionLifetimeHandler.OnTestSessionFinishingAsync(testSessionContext).ConfigureAwait(false);
-            }
-        }
-
-        if (hasNonDataConsumers)
-        {
-            // At this point, we called all non-consumer handlers.
-            // Now, we want to make sure to drain the message bus before calling the consumer handlers.
-            // Messages produced by non-consumer handlers could be consumed by consumer handlers.
-            await baseMessageBus.DrainDataAsync().ConfigureAwait(false);
-        }
-
-        foreach (ITestSessionLifetimeHandler testSessionLifetimeHandler in testSessionLifetimeHandlersContainer.TestSessionLifetimeHandlers)
-        {
-            if (testSessionLifetimeHandler is not IDataConsumer)
-            {
-                // We already called this for non-consumers.
-                continue;
-            }
-
-            using (otelService?.StartActivity(testSessionLifetimeHandler.Uid, testSessionLifetimeHandler.ToOTelTags()))
-            using (shutdownProgressReporter?.Track(testSessionLifetimeHandler.Uid, testSessionLifetimeHandler.DisplayName, nameof(ITestSessionLifetimeHandler.OnTestSessionFinishingAsync)))
-            {
-                await testSessionLifetimeHandler.OnTestSessionFinishingAsync(testSessionContext).ConfigureAwait(false);
-            }
-
-            // OnTestSessionFinishingAsync could produce information that needs to be handled by others.
-            // While in many cases we already handled this by calling all non-consumers first, it's possible that
-            // two consumers might depend on each other. In that case, we solely rely on registration order.
-            // And we drain in between.
-            await baseMessageBus.DrainDataAsync().ConfigureAwait(false);
-        }
-
-        // We disable after the drain because it's possible that the drain will produce more messages
-        await baseMessageBus.DrainDataAsync().ConfigureAwait(false);
-        await baseMessageBus.DisableAsync().ConfigureAwait(false);
-    }
-
-    protected static async Task DisposeServiceProviderAsync(ServiceProvider serviceProvider, Func<object, bool>? filter = null, List<object>? alreadyDisposed = null, bool isProcessShutdown = false)
-    {
-        alreadyDisposed ??= [];
-
-        // Close the message bus handshake before disposing anything at all. Consumers are routinely registered
-        // as services in their own right, and several of them (the output device, the coverage accumulator) land
-        // in Services *before* the bus does, so disabling inline when the loop happens to reach the bus would
-        // already have disposed those consumers while their ConsumeAsync was still running.
-        foreach (object service in serviceProvider.Services)
-        {
-            if (service is not BaseMessageBus messageBus || (filter is not null && !filter(messageBus)))
-            {
-                continue;
-            }
-
-            await EnsureMessageBusDisabledAsync(messageBus, serviceProvider).ConfigureAwait(false);
-
-            // Disabling is bounded on an aborted run, so it can return with a consumer still inside
-            // ConsumeAsync. Recording those as already disposed is what keeps every loop below off them:
-            // disposing one now is precisely the race the handshake exists to prevent, so we leak it and let
-            // the process exit reclaim it instead.
-            foreach (IDataConsumer dataConsumer in messageBus.ConsumersStillRunning)
-            {
-                if (alreadyDisposed.Contains(dataConsumer))
-                {
-                    continue;
-                }
-
-                alreadyDisposed.Add(dataConsumer);
-                await LogShutdownWarningAsync(
-                    serviceProvider,
-                    $"Not disposing data consumer '{dataConsumer.Uid}' because it is still consuming after the shutdown timeout elapsed.").ConfigureAwait(false);
-            }
-        }
-
-        foreach (object service in serviceProvider.Services)
-        {
-            // Logger is the most special service and we dispose it manually as last one, we want to be able to
-            // collect logs till the end of the process.
-            if (service is FileLoggerProvider)
-            {
-                continue;
-            }
-
-            // The LoggerFactoryProxy owns the real ILoggerFactory and is disposed manually after the rest of
-            // the services so that providers registered through it (including bridges to
-            // Microsoft.Extensions.Logging) can flush at the very end of the process.
-            if (service is LoggerFactoryProxy)
-            {
-                continue;
-            }
-
-            // The ITestApplicationCancellationTokenSource contains the cancellation token and can be used by other services during the shutdown
-            // we will collect manually in the correct moment.
-            if (service is ITestApplicationCancellationTokenSource)
-            {
-                continue;
-            }
-
-            if (filter is not null && !filter(service))
-            {
-                continue;
-            }
-
-            // We need to ensure that we won't dispose special services till the shutdown
-#pragma warning disable CS0618 // Type or member is obsolete
-            if (!isProcessShutdown &&
-                service is ITelemetryCollector or
-                 ITestHostApplicationLifetime or
-                 IPushOnlyProtocol or
-                 IPlatformOpenTelemetryService or
-                 IOpenTelemetryProvider)
-            {
-                continue;
-            }
-#pragma warning restore CS0618 // Type or member is obsolete
-
-            if (!alreadyDisposed.Contains(service))
-            {
-                await DisposeHelper.DisposeAsync(service).ConfigureAwait(false);
-                alreadyDisposed.Add(service);
-            }
-
-            if (service is BaseMessageBus busWithConsumers)
-            {
-                foreach (IDataConsumer dataConsumer in busWithConsumers.DataConsumerServices)
-                {
-                    if (filter is not null && !filter(dataConsumer))
-                    {
-                        continue;
-                    }
-
-                    // Consumers still inside ConsumeAsync were recorded as already disposed by the pre-pass
-                    // above, so this check is what spares them here.
-                    if (!alreadyDisposed.Contains(dataConsumer))
-                    {
-                        await DisposeHelper.DisposeAsync(dataConsumer).ConfigureAwait(false);
-                        alreadyDisposed.Add(dataConsumer);
-                    }
-                }
-            }
-        }
-    }
 }
